backend/requirements.txt
# --- Web framework ---
fastapi==0.136.0
uvicorn[standard]==0.44.0
starlette==1.0.0
sse-starlette==3.3.4

# --- HTTP / networking ---
httpx==0.28.1
requests==2.33.1

# --- Data validation ---
pydantic==2.13.2
pydantic_core==2.46.2

# --- Config ---
python-dotenv==1.2.2

# --- Email (SendGrid) ---
sendgrid==6.12.5

# --- Google Sheets (OAuth) ---
gspread==6.2.1
google-auth==2.49.2
google-auth-oauthlib==1.3.1

# --- LLM / AI agents ---
langchain==1.2.15
langchain-core==1.3.0
langchain-openai==1.1.14
openai==2.32.0
tiktoken==0.12.0

# --- LangGraph ---
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langgraph-supervisor==0.0.31
langsmith==0.7.32

# --- Web scraping (Firecrawl) ---
firecrawl-py>=4.22.0

# --- Utilities ---
orjson==3.11.8
tenacity==9.1.4
PyYAML==6.0.3
